requirements.txt
amqp==1.4.9
anyjson==0.3.3
billiard==3.3.0.23
blessings==1.6
boto3==1.4.1
botocore==1.4.74
celery==3.1.7
Django==1.9
django-celery==3.1.17
django-redis-cache==1.7.1
docutils==0.12
futures==3.0.5
inception==0.0.3
inquirer==2.1.11
Jinja2==2.8
jmespath==0.9.0
kombu==3.0.37
MarkupSafe==0.23
mongoengine==0.10.6
multi-key-dict==2.0.3
MySQL-python==1.2.5
Naked==0.1.31
pbr==2.0.0
pycrypto==2.6.1
pymongo==3.3.1
python-consul==0.6.1
python-dateutil==2.6.0
python-jenkins==0.4.14
pytz==2016.7
PyYAML==3.12
readchar==0.7
redis==2.10.5
requests==2.11.1
s3transfer==0.1.9
shellescape==3.4.1
six==1.10.0
vine==1.1.3
